.\Question 2.py

+++ Question 2.py
@@ -5,7 +5,7 @@ while True:
     counter += 1
     avg = sum / counter
 
-    num = int(input(f" Please enter a number {counter} (avg: {avg}, sum: {sum}): "))
+    num = float(input(f" Please enter a number {counter} (avg: {avg}, sum: {sum}): "))
 
     if num < 0:
         print(f"Thank you. Goodbye")
